Update isolinux.cfg to use proper VESA mode.
isolinux.bin and ldlinux.c32 are from syslinux 5.01, since 5.10 and 6.x were giving me grief. For some reason the last commit for isolinux.bin said that it was updated to syslinux 6.01 (which it wasn't). Also included in this commit is ldlinux.c32 from syslinux 5.01, since I have no idea where ldlinux.c32 is from in the first place.