Update the install script to a working state.
Removed some functions to make it a bit simpler. Removed copying of /home over as to create a blank state when installing (otherwise if no files in /home existed, the script would exit). Also fixed an issue where the script would assume that we are using /dev/sda when writing to /etc/lilo.conf. Thanks again to Onti q for the help :)